* Vulkan fixes/improvements/updates:    * added preliminary Vulkan 1.2 / SPIR-V 1.5 support (note that there is no compiler support yet, so this is treated as Vulkan 1.1 for now)    * replaced "VK_LAYER_LUNARG_standard_validation" layer with "VK_LAYER_KHRONOS_validation" (this is now standard and necessary to actually work)    * vulkan_kernel: when specializing a kernel, set the stage info to "compute"    * vulkan_kernel: the fixed sampler set workaround/awkwardness is no longer needed and actively prohibited, i.e. we no longer need to write/update the fixed sampler set when setting function arguments (descriptor sets)  * misc virtual -> override replacements  * atomic_compat.hpp: include <atomic> when compiling with the vanilla host toolchain (need memory_order enums)  * aligned_ptr: made it std:: compatible  * CMake: when searching for the glcorearb.h header, allow a mesa prefix (this is necessary for mingw/msys2)

+++ compute/device/host_image.hpp
@@ -394,7 +394,7 @@ FLOOR_POP_WARNINGS()
 		template <bool is_lod_ = is_lod, bool is_bias_ = is_bias, enable_if_t<is_lod_ || is_bias_>* = nullptr>
 		floor_inline_always static constexpr uint32_t select_lod(const int32_t lod_i, const float lod_or_bias_f) {
 			return ::min(host_limits::max_mip_levels - 1u, (is_lod_float || is_bias ?
-															uint32_t(::max(0.0f, ::round(lod_or_bias_f))) :
+															uint32_t(::max(0.0f, ::roundf(lod_or_bias_f))) :
 															uint32_t(::max(0, lod_i))));
 		}
 		// clamps lod to [0, max #mip-levels - 1]
